feat: move string-related utils to separate file
